REAFACTOR: move FileQuery and configs ADD: Converter skeleton
