+++ src/main/java/mops/businesslogic/DirectoryServiceImpl.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import mops.businesslogic.exception.DatabaseException;
 import mops.businesslogic.exception.DeleteAccessPermissionException;
 import mops.businesslogic.exception.StorageLimitationException;
+import mops.businesslogic.query.FileQuery;
 import mops.exception.MopsException;
 import mops.persistence.DirectoryPermissionsRepository;
 import mops.persistence.DirectoryRepository;
