+++ src/main/java/mops/persistence/config/DatabaseConfig.java
@@ -1,4 +1,4 @@
-package mops.config;
+package mops.persistence.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jdbc.repository.config.EnableJdbcAuditing;
