+++ src/test/java/mops/businesslogic/DirectoryServiceTest.java
@@ -2,6 +2,7 @@ package mops.businesslogic;
 
 import mops.businesslogic.exception.ReadAccessPermissionException;
 import mops.businesslogic.exception.WriteAccessPermissionException;
+import mops.businesslogic.query.FileQuery;
 import mops.exception.MopsException;
 import mops.persistence.DirectoryPermissionsRepository;
 import mops.persistence.FileRepository;
