+++ src/test/java/mops/businesslogic/FileQueryTest.java
@@ -1,5 +1,6 @@
 package mops.businesslogic;
 
+import mops.businesslogic.query.FileQuery;
 import mops.persistence.file.FileInfo;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
